Optional different displays of sustained fire sprites
See Issue #42
Instructions:
BEFORE RUNNING THE GAME, enable the "Sus Fire Only On Selection" check
box on the InputOutput class attached to GameController. Then sprites
will only show when the unit is selected.
If you don't want to use that, leave as-is.

Changelog:
- Implemented visibility of sustained fire sprites being able to be
visible at all times, or only when the unit is selected.
- Changed the variable of the selected unit to a Unit not a Gameobject