AI Efficiency
I'm not sure how effective this was, but it was the most I could do
without smashing my brain to pieces on the desk.
Refactored the Algorithm class, including adding the ShortestPaths class
as storage.
Added large numbers of yield statements to spread the calculations
across multiple frames.
Changed the class execution order.