Hologram Rotation on Button Hover
Input Handler:
- Changed all of the hologram activations to use gameobjects so that the
holograms can be rotated
- Added calls so that InputOutput has a boolean showing whether there
are holograms active or not.

InputOutput:
- Altered activePartSys to be a GameObject List
- Created new hologram methods allowing active holograms to be rotated
to a specified facing, reset the facing or all holograms on the map, and
deactivating all the holograms on the map, if any are active.
- Added calls to new methods from ShowActionSequence

FacingButtonHover
- Created the script
- Wrote code to find the InputOutput from GameController ant then call
the method to face the active hologram in the direction the button
represents.

Interactible:
- Altered hologram changing loops to use gameobjects instead of the
particle systems, added calls to the new boolean values in InputOutput,

Project:
- Added the FacingButtonHover script to the facing canvas buttons
- Also added an EventSystem component to the buttons and used this to
call the faceHologram method in the FacingButtonHover script. (As
OnMouseOver does not work with buttons).